Make the vscode API globally accessible without the need to pass it to each individual component.
