--- src/VSCodeApi.ts
@@ -1,5 +1,0 @@
-export default interface VSCodeApi {
-  getState: () => any
-  setState: (newState: any) => any
-  postMessage: (message: any) => void
-}
